Fixed onaudioprocess finalizing
There is an edge case, when audio callback is still called after closing audio device, which leads to crash because audio device doesn't exists anymore. In this function there is a proper scriptProcessorNode finalizing code, but only for "if capture is enabled" branch, which is placed above.

So the fix is to use the same method for clearing onaudioprocess as in the code above.